Moving pragma/program to preprocessor
